70. Shading added to more distant walls. Floor and ceiling being colored.
Solution to room-traversing ideated:
at every move, simply check if the main direction ray intersects with the portal. Get its distance and see if it is less than given speed.
If the distance is less, need to change rooms:
In current room, get the room of correct direction.
Move the position, and wrap it around to th eother direction.
Remeber only do this if the vector passes through portal. Else, simply clamp the position inside the current room.